DOMCore/tests/submissions/Ms2ger/tests/Document-createProcessingInstruction-literal-2.xhtml

<?xml-stylesheet href="support/style.css" type="text/css"?>
<html xmlns="http://www.w3.org/1999/xhtml">
<head>
<title>Document.createProcessingInstruction</title>
<script src="testharness.js"/>
</head>
<body>
<div id="log"/>
<script>
test(function() {
  var pienc = document.firstChild;
  assert_true(pienc instanceof ProcessingInstruction)
  assert_equals(pienc.target, "xml-stylesheet")
  assert_equals(pienc.data, 'href="support/style.css" type="text/css"')
})
</script>
</body>
</html>
